requirements.txt
videosdk==0.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
